Keep the command for generating the comic the same as before
